[BACKPORT][DoNotCarryForward] Clipboard: move various predefined FormatTypes to common file.
Many clipboard implementations use the predefined FormatTypes
from the clipboard_aurax11. Thus, to make clipboard ozone use
them as well, move those into a separate clipboard_linux
file and include it when necessary.
