+++ src/ui/base/BUILD.gn
@@ -650,6 +650,7 @@ component("base") {
     # Aura clipboard.
     if (use_x11 && is_desktop_linux) {
       sources += [
+        "clipboard/clipboard_aura_linux.cc",
         "clipboard/clipboard_aurax11.cc",
         "clipboard/clipboard_aurax11.h",
       ]
