+++ src/ui/base/clipboard/clipboard.h
@@ -16,6 +16,7 @@
 #include "base/containers/flat_map.h"
 #include "base/lazy_instance.h"
 #include "base/macros.h"
+#include "base/no_destructor.h"
 #include "base/process/process.h"
 #include "base/strings/string16.h"
 #include "base/synchronization/lock.h"
@@ -83,6 +84,7 @@ class UI_BASE_EXPORT Clipboard : public base::ThreadChecker {
 
    private:
     friend class Clipboard;
+    friend class base::NoDestructor<FormatType>;
 
     // Platform-specific glue used internally by the Clipboard class. Each
     // plaform should define,at least one of each of the following:
